Refactor, improve and better test ASTER functionality.
This commit modifies the aster subpackage to break apart the creation
of the cogs from the creation of the stac items. It also bases the
stac item creation mostly on the XML metadata. It drops a lot of
noisey metadata an generates more pertinent STAC metadata, like
additional assets and proj information.